refactor(userspace): extract per-feature status counters to protocol_counters.go
Eighth code-motion increment of the protocol.go wire-protocol split
(#4839). Moves the per-feature runtime status and hit-counter types
reported inside ProcessStatus into a cohesive sibling
protocol_counters.go in the same userspace package.

Moved verbatim (no rename, no logic or json-tag change):
SourceNATPoolStatus, CoSInterfaceStatus, ThreeColorPolicerStatus,
CoSQueueStatus, CoSActiveFlowCountStatus, FirewallFilterTermCounterStatus,
PolicyRuleCounterStatus, NATRuleCounterStatus, ZoneTrafficCounterStatus.

These mirror the Rust counter types in userspace-dp/src/protocol; every
serde-matching json tag is preserved byte-identical. Build passes;
gofmt clean.
